update AKS config
