+++ compute/container_cluster.go
@@ -1,4 +1,4 @@
-package containerservice
+package compute
 
 import (
 	"context"
@@ -8,20 +8,18 @@ import (
 	"os"
 	"time"
 
-	"github.com/Azure-Samples/azure-sdk-for-go-samples/helpers"
-	"github.com/Azure-Samples/azure-sdk-for-go-samples/iam"
 	"github.com/Azure/azure-sdk-for-go/services/containerservice/mgmt/2017-09-30/containerservice"
+
+	"github.com/Azure-Samples/azure-sdk-for-go-samples/internal/config"
+	"github.com/Azure-Samples/azure-sdk-for-go-samples/internal/iam"
 	"github.com/Azure/go-autorest/autorest/to"
 )
 
-// fakepubkey is used if a key isn't available at the specified path in CreateAKS(...)
-var fakepubkey = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC7laRyN4B3YZmVrDEZLZoIuUA72pQ0DpGuZBZWykCofIfCPrFZAJgFvonKGgKJl6FGKIunkZL9Us/mV4ZPkZhBlE7uX83AAf5i9Q8FmKpotzmaxN10/1mcnEE7pFvLoSkwqrQSkrrgSm8zaJ3g91giXSbtqvSIj/vk2f05stYmLfhAwNo3Oh27ugCakCoVeuCrZkvHMaJgcYrIGCuFo6q0Pfk9rsZyriIqEa9AtiUOtViInVYdby7y71wcbl0AbbCZsTSqnSoVxm2tRkOsXV6+8X4SnwcmZbao3H+zfO1GBhQOLxJ4NQbzAa8IJh810rYARNLptgmsd4cYXVOSosTX azureuser"
-
 func getAKSClient() (containerservice.ManagedClustersClient, error) {
-	aksClient := containerservice.NewManagedClustersClient(helpers.SubscriptionID())
-	auth, _ := iam.GetResourceManagementAuthorizer(iam.AuthGrantType())
+	aksClient := containerservice.NewManagedClustersClient(config.SubscriptionID())
+	auth, _ := iam.GetResourceManagementAuthorizer()
 	aksClient.Authorizer = auth
-	aksClient.AddToUserAgent(helpers.UserAgent())
+	aksClient.AddToUserAgent(config.UserAgent())
 	aksClient.PollingDuration = time.Hour * 1
 	return aksClient, nil
 }
