+++ compute/vm.go
@@ -20,13 +20,6 @@ import (
 	"github.com/Azure/go-autorest/autorest/to"
 )
 
-const (
-	publisher  = "Canonical"
-	offer      = "UbuntuServer"
-	sku        = "16.04.0-LTS"
-	fakepubkey = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC7laRyN4B3YZmVrDEZLZoIuUA72pQ0DpGuZBZWykCofIfCPrFZAJgFvonKGgKJl6FGKIunkZL9Us/mV4ZPkZhBlE7uX83AAf5i9Q8FmKpotzmaxN10/1mcnEE7pFvLoSkwqrQSkrrgSm8zaJ3g91giXSbtqvSIj/vk2f05stYmLfhAwNo3Oh27ugCakCoVeuCrZkvHMaJgcYrIGCuFo6q0Pfk9rsZyriIqEa9AtiUOtViInVYdby7y71wcbl0AbbCZsTSqnSoVxm2tRkOsXV6+8X4SnwcmZbao3H+zfO1GBhQOLxJ4NQbzAa8IJh810rYARNLptgmsd4cYXVOSosTX azureuser"
-)
-
 func getVMClient() compute.VirtualMachinesClient {
 	vmClient := compute.NewVirtualMachinesClient(config.SubscriptionID())
 	a, _ := iam.GetResourceManagementAuthorizer()
